+++ lib/types.ts
@@ -1,0 +1,7 @@
+export type Product = {
+  id: string;
+  name: string;
+  price: number;
+  description: string;
+  imageUrl: string;
+};
